+++ packages/fiori-app-sub-generator/test/int/fiori-elements/expected-output/lrop_v4_cap/app/lrop_v4_cap/webapp/test/integration/pages/JourneyRunner.js
@@ -1,21 +1,19 @@
-sap.ui.define(
-    [
-        'sap/fe/test/JourneyRunner',
-        'testNameSpace/lropv4cap/test/integration/pages/BooksList',
-        'testNameSpace/lropv4cap/test/integration/pages/BooksObjectPage'
-    ],
-    function (JourneyRunner, BooksList, BooksObjectPage) {
-        'use strict';
+sap.ui.define([
+    "sap/fe/test/JourneyRunner",
+	"testNameSpace/lropv4cap/test/integration/pages/BooksList",
+	"testNameSpace/lropv4cap/test/integration/pages/BooksObjectPage"
+], function (JourneyRunner, BooksList, BooksObjectPage) {
+    'use strict';
 
-        var runner = new JourneyRunner({
-            launchUrl: sap.ui.require.toUrl('testNameSpace/lropv4cap') + '/test/flp.html#app-preview',
-            pages: {
-                onTheBooksList: BooksList,
-                onTheBooksObjectPage: BooksObjectPage
-            },
-            async: true
-        });
+    var runner = new JourneyRunner({
+        launchUrl: sap.ui.require.toUrl('testNameSpace/lropv4cap') + '/test/flp.html#app-preview',
+        pages: {
+			onTheBooksList: BooksList,
+			onTheBooksObjectPage: BooksObjectPage
+        },
+        async: true
+    });
 
-        return runner;
-    }
-);
+    return runner;
+});
+