+++ packages/abap-deploy-config-sub-generator/CHANGELOG.md
@@ -1,5 +1,32 @@
 # @sap-ux/abap-deploy-config-sub-generator
 
+## 1.0.11
+
+### Patch Changes
+
+- Updated dependencies [c8e8f7e]
+    - @sap-ux/ui5-config@1.0.3
+    - @sap-ux/abap-deploy-config-writer@1.0.7
+    - @sap-ux/adp-tooling@1.0.11
+    - @sap-ux/abap-deploy-config-inquirer@2.0.9
+    - @sap-ux/project-access@2.1.2
+    - @sap-ux/axios-extension@2.0.2
+    - @sap-ux/fiori-generator-shared@1.0.9
+    - @sap-ux/deploy-config-generator-shared@1.0.9
+
+## 1.0.10
+
+### Patch Changes
+
+- Updated dependencies [0fa8305]
+    - @sap-ux/btp-utils@2.0.2
+    - @sap-ux/abap-deploy-config-inquirer@2.0.8
+    - @sap-ux/adp-tooling@1.0.10
+    - @sap-ux/axios-extension@2.0.2
+    - @sap-ux/deploy-config-generator-shared@1.0.8
+    - @sap-ux/fiori-generator-shared@1.0.8
+    - @sap-ux/abap-deploy-config-writer@1.0.6
+
 ## 1.0.9
 
 ### Patch Changes